Pass around h2o_jar. Add signal handler.

+++ R/tests/run.py
@@ -3,6 +3,8 @@
 import sys
 import os
 import shutil
+import signal
+import time
 
 
 class H2OCloudNode:
@@ -12,13 +14,15 @@ class H2OCloudNode:
     port: The actual port chosen at run time.
     pid: The process id of the node.
     """
-    def __init__(self, cloud_num, node_num, base_port, xmx, output_dir):
+    def __init__(self, cloud_num, node_num, h2o_jar, base_port, xmx, output_dir):
         self.cloud_num = cloud_num
         self.node_num = node_num
+        self.h2o_jar = h2o_jar
         self.base_port = base_port
         self.xmx = xmx
         self.output_dir = output_dir
 
+        self.terminated = False
         self.port = -1
         self.pid = -1
 
@@ -27,6 +31,12 @@ class H2OCloudNode:
 
     def stop(self):
         pass
+
+    def terminate(self):
+        self.terminated = True
+        if (self.pid > 0):
+            print("Killing JVM with PID {}".format(self.pid))
+            os.kill(self.pid)
 
     def __str__(self):
         s = ""
@@ -41,18 +51,20 @@ class H2OCloud:
     """
     A class representing one of the H2O clouds.
     """
-    def __init__(self, cloud_num, nodes_per_cloud, base_port, xmx, output_dir):
+    def __init__(self, cloud_num, nodes_per_cloud, h2o_jar, base_port, xmx, output_dir):
         self.cloud_num = cloud_num
         self.nodes_per_cloud = nodes_per_cloud
+        self.h2o_jar = h2o_jar
         self.base_port = base_port
         self.xmx = xmx
         self.output_dir = output_dir
 
+        self.terminated = False
         self.nodes = []
         self.jobs_run = 0
 
         for i in range(self.nodes_per_cloud):
-            node = H2OCloudNode(self.cloud_num, i, self.base_port, self.xmx, self.output_dir)
+            node = H2OCloudNode(self.cloud_num, i, h2o_jar, self.base_port, self.xmx, self.output_dir)
             self.nodes.append(node)
 
     def start(self):
@@ -60,6 +72,11 @@ class H2OCloud:
 
     def stop(self):
         pass
+
+    def terminate(self):
+        self.terminated = True
+        for node in self.nodes:
+            node.terminate()
 
     def __str__(self):
         s = ""
@@ -70,18 +87,46 @@ class H2OCloud:
         return s
 
 
+class Test:
+    """
+    A class representing one Test.
+    """
+
+    def __init__(self, test_dir, test_name, output_dir):
+        self.test_dir = test_dir
+        self.test_name = test_name
+        self.output_dir = output_dir
+
+        self.terminated = False
+        self.running = False
+        self.exit_status = -9999
+        self.pid = -1
+
+    def is_running(self):
+        return self.running
+
+    def start(self):
+        pass
+
+    def terminate(self):
+        if (self.pid > 0):
+            print("Killing Test with PID {}".format(self.pid))
+            os.kill(self.pid)
+
+
 class RUnitRunner:
     """
     A class for running the RUnit tests.
     """
 
-    def __init__(self, test_root_dir, num_clouds, nodes_per_cloud, base_port, xmx, output_dir):
+    def __init__(self, test_root_dir, num_clouds, nodes_per_cloud, h2o_jar, base_port, xmx, output_dir):
         """
         Constructor.
 
         @param test_root_dir:
         @param num_clouds: Number of H2O clouds to start.
         @param nodes_per_cloud: Number of H2O nodes to start per cloud.
+        @param h2o_jar: Path to H2O jar file to run.
         @param base_port: Base H2O port (e.g. 54321) to start choosing from.
         @param xmx: Java -Xmx parameter.
         @param output_dir: Directory for output files.
@@ -90,14 +135,17 @@ class RUnitRunner:
         self.test_root_dir = test_root_dir
         self.num_clouds = num_clouds
         self.nodes_per_cloud = nodes_per_cloud
+        self.h2o_jar = h2o_jar
         self.base_port = base_port
         self.output_dir = output_dir
 
+        self.terminated = False
         self.clouds = []
         self.tests = []
+        self.tests_not_started = []
 
         for i in range(self.num_clouds):
-            cloud = H2OCloud(i, self.nodes_per_cloud, self.base_port, xmx, self.output_dir)
+            cloud = H2OCloud(i, self.nodes_per_cloud, h2o_jar, self.base_port, xmx, self.output_dir)
             self.clouds.append(cloud)
 
     def run(self):
@@ -106,11 +154,29 @@ class RUnitRunner:
 
         @return: none
         """
-        self.create_output_dir()
-        self.build_test_list()
-        self.start_clouds()
-        self.run_tests()
+        if (not self.terminated):
+            self.create_output_dir()
+
+        if (not self.terminated):
+            self.build_test_list()
+
+        if (not self.terminated):
+            self.start_clouds()
+
+        if (not self.terminated):
+            self.run_tests()
+
         self.stop_clouds()
+
+    def terminate(self):
+        self.terminated = True
+
+        for test in self.tests:
+            if (test.is_running()):
+                test.terminate()
+
+        for cloud in self.clouds:
+            cloud.terminate()
 
     #--------------------------------------------------------------------
     # Private methods below this line.
@@ -118,7 +184,7 @@ class RUnitRunner:
 
     def create_output_dir(self):
         try:
-            os.mkdir(self.output_dir)
+            os.makedirs(self.output_dir)
         except OSError as e:
             print("")
             print("mkdir failed (errno {0}): {1}".format(e.errno, e.strerror))
@@ -148,6 +214,7 @@ class RUnitRunner:
         s = "\n"
         s += "test_root_dir:    {}\n".format(self.test_root_dir)
         s += "output_dir:       {}\n".format(self.output_dir)
+        s += "h2o_jar:          {}\n".format(self.h2o_jar)
         s += "num_clouds:       {}\n".format(self.num_clouds)
         s += "nodes_per_cloud:  {}\n".format(self.nodes_per_cloud)
         s += "base_port:        {}\n".format(self.base_port)
@@ -165,6 +232,27 @@ base_port = 40000
 num_clouds = 3
 wipe_output_dir = False
 output_dir = ""
+h2o_jar = ""
+runner = None
+handling_signal = False
+
+
+def signal_handler(signum, stackframe):
+    global runner
+    global handling_signal
+
+    if (handling_signal):
+        # Don't do this recursively.
+        return
+    handling_signal = True
+
+    print("")
+    print("----------------------------------------------------------------------")
+    print("")
+    print("SIGNAL CAUGHT.  SHUTTING DOWN NOW.")
+    print("")
+    print("----------------------------------------------------------------------")
+    runner.terminate()
 
 
 def usage():
@@ -205,14 +293,25 @@ def parse_args(argv):
 
 
 def main(argv):
+    global runner
+
     test_root_dir = os.path.dirname(os.path.realpath(__file__))
     output_dir = test_root_dir + "/" + "results"
+    h2o_jar = os.path.abspath(test_root_dir + "/../../target/h2o.jar")
 
     parse_args(argv)
 
+    if (not os.path.exists(h2o_jar)):
+        print("")
+        print("h2o jar not found: {}".format(h2o_jar))
+        print("    " + output_dir)
+        print("")
+        sys.exit(1)
+
     if (wipe_output_dir):
         try:
-            shutil.rmtree(output_dir)
+            if (os.path.exists(output_dir)):
+                shutil.rmtree(output_dir)
         except OSError as e:
             print("")
             print("removing directory failed (errno {0}): {1}".format(e.errno, e.strerror))
@@ -222,10 +321,12 @@ def main(argv):
 
     nodes_per_cloud = 1
     xmx = "2g"
-    runner = RUnitRunner(test_root_dir, num_clouds, nodes_per_cloud, base_port, xmx, output_dir)
+    runner = RUnitRunner(test_root_dir, num_clouds, nodes_per_cloud, h2o_jar, base_port, xmx, output_dir)
+    signal.signal(signal.SIGINT, signal_handler)
+    signal.signal(signal.SIGTERM, signal_handler)
     runner.run()
     print str(runner)
 
-
 if __name__ == "__main__":
     main(sys.argv)
+    time.sleep(100)
